+++ archiso/profiledef.sh
@@ -1,10 +1,10 @@
 #!/usr/bin/env bash
 # shellcheck disable=SC2034
 
-iso_name="enigmaos"
-iso_label="ENIGMAOS_$(date --date="@${SOURCE_DATE_EPOCH:-$(date +%s)}" +%Y%m)"
-iso_publisher="EnigmaOS <https://enigmaos.rishispace.dev>"
-iso_application="EnigmaOS Live/Install Medium"
+iso_name="enigmarsos"
+iso_label="ENIGMARSOS_$(date --date="@${SOURCE_DATE_EPOCH:-$(date +%s)}" +%Y%m)"
+iso_publisher="EnigmarsOS <https://enigmarsos.rishispace.dev>"
+iso_application="EnigmarsOS Live/Install Medium"
 iso_version="$(date --date="@${SOURCE_DATE_EPOCH:-$(date +%s)}" +%Y.%m.%d)"
 install_dir="arch"
 buildmodes=('iso')
@@ -16,12 +16,12 @@ airootfs_image_type="squashfs"
 airootfs_image_tool_options=('-comp' 'xz' '-Xbcj' 'x86' '-b' '1M' '-Xdict-size' '1M')
 bootstrap_tarball_compression=('zstd' '-c' '-T0' '--auto-threads=logical' '--long' '-19')
 file_permissions=(
-  ["/usr/share/enigmaos/scripts/install-limine.sh"]="0:0:755"
-  ["/usr/share/enigmaos/scripts/seed-kernel.sh"]="0:0:755"
-  ["/usr/share/enigmaos/scripts/fix-mkinitcpio.sh"]="0:0:755"
-  ["/usr/share/enigmaos/scripts/post-install.sh"]="0:0:755"
-  ["/usr/local/bin/enigmaos-branding"]="0:0:755"
-  ["/usr/share/libalpm/scripts/enigmaos-os-release"]="0:0:755"
+  ["/usr/share/enigmarsos/scripts/install-limine.sh"]="0:0:755"
+  ["/usr/share/enigmarsos/scripts/seed-kernel.sh"]="0:0:755"
+  ["/usr/share/enigmarsos/scripts/fix-mkinitcpio.sh"]="0:0:755"
+  ["/usr/share/enigmarsos/scripts/post-install.sh"]="0:0:755"
+  ["/usr/local/bin/enigmarsos-branding"]="0:0:755"
+  ["/usr/share/libalpm/scripts/enigmarsos-os-release"]="0:0:755"
   ["/etc/shadow"]="0:0:400"
   ["/root"]="0:0:750"
   ["/root/.automated_script.sh"]="0:0:755"
@@ -29,7 +29,7 @@ file_permissions=(
   ["/usr/local/bin/choose-mirror"]="0:0:755"
   ["/usr/local/bin/Installation_guide"]="0:0:755"
   ["/usr/local/bin/livecd-sound"]="0:0:755"
-  ["/usr/local/bin/enigmaos-welcome"]="0:0:755"
-  ["/usr/local/bin/enigmaos-firstboot"]="0:0:755"
-  ["/usr/local/bin/enigmaos-setup-live-user"]="0:0:755"
+  ["/usr/local/bin/enigmarsos-welcome"]="0:0:755"
+  ["/usr/local/bin/enigmarsos-firstboot"]="0:0:755"
+  ["/usr/local/bin/enigmarsos-setup-live-user"]="0:0:755"
 )